+++ performance/suites/res-send.js
@@ -3,6 +3,7 @@
  */
 const { report } = require('./util')
 const Benchmark = require('benchmark')
+const EventEmitter = require('events')
 const suite = new Benchmark.Suite()
 
 const extensions = require('../../libs/response-extensions')
@@ -16,11 +17,6 @@ const send = extensions.send({}, {}, {
 
 const buffer = Buffer.from('Hello World!')
 const json = { msg: 'Hello World' }
-const stream = {
-  pipe () {},
-  once () {},
-  unpipe () {}
-}
 const headers = {
   'content-type': 'text/plain',
   'x-framework': 'restana',
@@ -58,8 +54,18 @@ suite
   .add('buffer', function () {
     send(buffer)
   })
-  .add('stream', function () {
-    send(stream)
+  .add('stream setup', function () {
+    const source = new EventEmitter()
+    source.pipe = () => {}
+    source.unpipe = () => {}
+    const response = new EventEmitter()
+    response.statusCode = 200
+    response.end = () => {}
+    response.setHeader = () => {}
+    response.getHeader = () => {}
+    response.removeHeader = () => {}
+    response.destroy = () => {}
+    extensions.send({ errorHandler () {} }, {}, response)(source)
   })
   .on('complete', function () {
     report(this)